requirements.txt
Django==1.9
PuLP==1.5.9
PyJWT==1.1.0
beautifulsoup4==4.3.2
dj-database-url==0.3.0
dj-static==0.0.6
django-toolbelt==0.0.1
djangorestframework==3.1.1
gunicorn==19.3.0
numpy==1.9.1
oauthlib==0.7.2
psycopg2==2.6
pyparsing==1.5.7
python-openid==2.2.5
python-social-auth==0.2.7
requests==2.7.0
requests-oauthlib==0.5.0
scipy==0.14.0
six==1.9.0
static3==0.5.1
wsgiref==0.1.2
